Renamed the facade to UrlSigner and moved to a Facade namespace
